pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "homepilot"
version = "0.1.0"
description = "Home Lab Manager TUI — manage, monitor, and deploy apps across TrueNAS, Proxmox, and more"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Neil" },
]
dependencies = [
    "textual>=0.47.0",
    "paramiko>=3.4.0",
    "pyyaml>=6.0.1",
    "click>=8.1.7",
    "httpx>=0.27.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "textual-dev>=1.5.0",
]

[project.scripts]
homepilot = "homepilot.__main__:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/homepilot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
